requirements.txt
# Dipendenze Python
fastapi 
uvicorn 
python-jose[cryptography] 
passlib[bcrypt] 
mysql-connector-python
os
dotenv

